+++ src/Event.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Cspray\Labrador\AsyncEvent;
+namespace Labrador\AsyncEvent;
 
 use DateTimeImmutable;
 use DateTimeInterface;
@@ -8,12 +8,13 @@ use DateTimeInterface;
 /**
  * Implementations that represent a semantic event triggered by Labrador or a Labrador powered Application.
  *
- * @package Cspray\Labrador\AsyncEvent
+ * @package Labrador\AsyncEvent
  */
 interface Event {
 
     /**
-     * The name of the Event, listeners who register for this event must match this string to be invoked.
+     * The name of the Event being invoked; this value is passed to Listener::canHandle to determine if the given
+     * Listener should be invoked.
      *
      * @return string
      */
@@ -23,9 +24,8 @@ interface Event {
      * An object that represents the target of the event.
      *
      * In the context of Labrador events MUST be invoked either on or in context of an object. Generally the target
-     * of the event should be obvious. For example, the target of the Engine::START_UP_EVENT in Labrador\Core is the
-     * Engine implementation. If your targets are not intrinsically targeting an object it may be a sign that your
-     * types are not fully designed.
+     * of the event should be apparent. If your targets are not intrinsically targeting an object it may be a sign that
+     * your types are not fully designed.
      *
      * @return object
      */
@@ -34,7 +34,7 @@ interface Event {
     /**
      * An arbitrary set of data associated to this Event.
      *
-     * Perhaps the target alone is not sufficient data to pass about this event or there is meta data that you'd like
+     * Perhaps the target alone is not sufficient data to pass about this event or there is meta-data that you'd like
      * to associate with the event (e.g. what host triggered an event in a network-enabled emitter).
      *
      * @return array