+++ test/Integration/AutowiringListenersTest.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Cspray\Labrador\AsyncEvent\Test\Integration;
+namespace Labrador\AsyncEvent\Test\Integration;
 
 use Amp\PHPUnit\AsyncTestCase;
 use Cspray\AnnotatedContainer\AnnotatedContainer;
 use Cspray\AnnotatedContainer\Bootstrap;
-use Cspray\Labrador\AsyncEvent\AmpEventEmitter;
-use Cspray\Labrador\AsyncEvent\EventEmitter;
-use Cspray\Labrador\AsyncEvent\StandardEvent;
+use Labrador\AsyncEvent\AmpEventEmitter;
+use Labrador\AsyncEvent\EventEmitter;
+use Labrador\AsyncEvent\StandardEvent;
 
 class AutowiringListenersTest extends AsyncTestCase {