+++ test/Integration/Listeners/FooListener.php
@@ -1,14 +1,14 @@
 <?php
 
-namespace Cspray\Labrador\AsyncEvent\Test\Integration\Listeners;
+namespace Labrador\AsyncEvent\Test\Integration\Listeners;
 
 use Amp\Future;
-use Cspray\Labrador\AsyncEvent\AbstractListener;
-use Cspray\Labrador\AsyncEvent\DependencyInjection\AutowiredListener;
-use Cspray\Labrador\AsyncEvent\Event;
+use Labrador\AsyncEvent\AbstractListener;
+use Labrador\AsyncEvent\EventListener;
+use Labrador\AsyncEvent\Event;
 use Labrador\CompositeFuture\CompositeFuture;
 
-#[AutowiredListener]
+#[EventListener]
 class FooListener extends AbstractListener {
 
     public function handle(Event $event) : Future|CompositeFuture|null {